refactor(acrawl-cli): extract classic_repl.rs and resume.rs
